backend/requirements.txt
fastapi==0.135.2
uvicorn[standard]==0.42.0

SQLAlchemy==2.0.44
psycopg[binary]==3.3.3
alembic==1.15.1

python-jose==3.4.0
PyJWT==2.10.1
passlib[bcrypt]==1.7.4

pydantic==2.12.5
pydantic-settings==2.13.1
email-validator==2.2.0

python-multipart==0.0.20

pypdf==5.3.0
docx2txt==0.8

httpx==0.28.1

sentence-transformers>=3.0.0
openai>=1.30.0
tiktoken>=0.7.0
celery>=5.4.0
redis>=5.0.0
boto3>=1.34.0
sentry-sdk>=2.20.0
